pyproject.toml
[project]
name = "src"
version = "0.0.1"
dependencies = [
    "django==4.2",
    "pillow",
    "django-environ",
    "django-split-settings",
    "bcrypt",
    "django-behaviors",
    "redis",
    "celery",
    # "djangorestframework",  # djangorestframework-camel-case
    # "djangorestframework-simplejwt",  # drf-jwt
    # "django-filter",

    "django-health-check",  # https://github.com/revsys/django-health-check/tree/master
    # "drf-spectacular-sidecar",
    # "sentry-sdk",
    "psycopg2-binary",

    # "django-ipware",
    # "django-axes",
    # "whitenoise",
]

[project.optional-dependencies]
dev = [
    "ipython",
    "black",
    "isort",
    "autoflake",
    "django-debug-toolbar",
    # "dotenv-linter",

    "faker",
    "mixer",
    "pytest-factoryboy",
    "pytest-django",
    "pytest-cov",
    "pytest-deadfixtures",
    "pytest-randomly",
    "pytest-xdist",
    # "pytest-env",
    # "pytest-freezegun",
    # "pytest-mock",
    # "pytest-repeat",
    # "pytest-httpx",

    # "mypy",

    # "flake8",
]

[tool.black]
 exclude = '''
 /(
   | migrations
 )/
 '''
 line_length = 120

[tool.isort]
profile = "google"
line_length = 120
lines_after_imports=2
multi_line_output = 3
use_parentheses = true
extra_standard_library = ["pytest"]
known_django = ["django"]  # "restframework"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
addopts = "--reuse-db"

filterwarnings = [
  "ignore::django.utils.deprecation.RemovedInDjango50Warning:django.conf",
  "ignore::django.core.paginator.UnorderedObjectListWarning",
]

python_files = "test*.py"